fix the linking error when USER_LDFLAGS contains "-Wl,--as-needed".
When building a shared library, the needed library comes after --as-needed
will be ignored by gcc as shared libraried are allowed to have undefined
symbols, so actually libproxychains4.so does not link against libpthread
and libdl.

This problem occurs when --as-no-needed in LDFLAGS is overwritten by --as-needed
in USER_LDFLAGS, we should place LDFLAGS(defined in Makefile) after USER_LDFLAGS
to make sure that --as-no-needed is working.