Suppress ‘function types not truly compatible in ISO C’ warnings.
These warnings are unavoidable when we redefine connect() and sendto()
since the GNU C library implements them using transparent unions as
their ‘sockaddr’ parameter, which we're not really supposed to be aware
of.

Implementing our own versions with transparent unions only introduces
more incompatibility with ISO C warnings, so we resort to using pragmas
to suppress them explicitly.